@Formula 구문 오류로 중간점검용 커밋

+++ src/main/kotlin/com/gcc/miti/module/dto/makeGroupDto/GroupDto.kt
@@ -7,7 +7,7 @@ import java.time.LocalDateTime
 data class GroupDto(
     val description: String,
     val title: String,
-    val maxUsers: Short,
+    val maxUsers: Long,
     val meetDate: LocalDateTime?,
     val meetPlace: String?,
 ) {
